Not an issue but an idea I had when I saw the greeting message "What a beautiful day to run a lake model."!
Since it appears that the tool has vocation to be educational too, you could have different greeting messages coming in.

If created in your script a table "greet", from which a row is sampled anytime you run a model, and send back a different message.

The messages I put are a bit corny, but this is to explain the idea.

Feel free to disregard, obviously :D